training/nnue/requirements.txt
numpy>=1.24
python-chess>=1.999
PyYAML>=6.0
torch>=2.0
tqdm>=4.66
